Restructure directories in prearation for adding mobile
