+++ client/package.json
@@ -18,6 +18,7 @@
   "homepage": "https://github.com/znrm/typedraw#readme",
   "dependencies": {
     "axios": "^0.18.0",
+    "lodash": "^4.17.10",
     "lodash.merge": "^4.6.1",
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
