+++ client/shared/reducers/entities/entities_reducer.js
@@ -3,5 +3,5 @@ import { combineReducers } from 'redux';
 import users from './users_reducer';
 
 export default combineReducers({
-    users
+  users
 });
