+++ client/shared/reducers/entities/users_reducer.js
@@ -6,13 +6,14 @@ import { START_SESSION } from '../../actions/session_actions';
 
 const UsersReducer = (oldState = {}, action) => {
   Object.freeze(oldState);
-  let newState = {};
+  const newState = {};
 
   switch (action.type) {
     case START_SESSION:
-    case RECEIVE_USER:
+    case RECEIVE_USER: {
       const userId = action.res.data._id;
       return merge(newState, oldState, { [userId]: action.res.data });
+    }
     default:
       return oldState;
   }
