+++ client/shared/reducers/error/errors_reducer.js
