+++ client/shared/reducers/session/session_reducer.js
