+++ client/shared/store/initialStore.js
@@ -1,0 +1,5 @@
+const initialStore = {
+
+};
+
+export default initialStore;
