+++ client/shared/util/session_api_util.js
@@ -1,0 +1,10 @@
+const axios = require('axios');
+
+export const fetchCurrentSession = () =>
+  axios.get('http://localhost:5000/api/session');
+
+export const startSession = user =>
+  axios.post('http://localhost:5000/api/session/login', user);
+
+export const endSession = () =>
+  axios.delete('http://localhost:5000/api/session');
