+++ client/shared/util/user_api_util.js
@@ -1,8 +1,7 @@
 const axios = require('axios');
 
-export const createUser = ({ email, password }) => (
+export const createUser = ({ email, password }) =>
   axios.post('http://localhost:5000/api/users/register', {
     email,
     password
-  })
-);
+  });
