+++ client/web/components/session_form/login_form_container.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
-import { login } from '../../actions/session_actions';
-import { removeErrors } from '../../actions/errors_actions';
+import { login } from '../../../shared/actions/session_actions';
+import { removeErrors } from '../../../shared/actions/errors_actions';
 import Form from './form';
 
 const mapStateToProps = ({ errors, session }) => ({
