--- client/web/store.js
@@ -1,17 +1,0 @@
-import { createStore, applyMiddleware } from 'redux';
-import RootReducer from './reducers/root_reducer';
-import thunk from 'redux-thunk';
-import logger from 'redux-logger';
-
-const configureStore = (preloadedState = {}) => (
-    createStore(
-        RootReducer,
-        preloadedState,
-        applyMiddleware(
-            thunk,
-            logger
-        )
-    )
-);
-
-export default configureStore;
