+++ server/config/passport.js
@@ -4,7 +4,7 @@ const mongoose = require('mongoose');
 require('../models/User');
 
 const User = mongoose.model('users');
-const keys = { secretOrKey: process.env.SECRET_OR_KEY } || require('../config/keys');
+const keys = { secretOrKey: process.env.SECRET_OR_KEY || require('../config/keys').secretOrKey };
 
 const options = {};
 options.jwtFromRequest = ExtractJwt.fromAuthHeaderAsBearerToken();
